# docs/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "PoplarGrid API 服务器，提供项目、成员等操作功能",
        "title": "PoplarGrid 核心 API 服务器",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "paths": {
...
        "/projects": {
            "get": {
                "description": "根据作品集 ID 获取项目列表，支持分页、排序和状态筛选。当列表为空时，会返回 null 而不是空数组。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project"
                ],
                "summary": "获取项目列表分页",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码，默认值为 1",
                        "name": "page_serial",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量，默认值为 10",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "排序方式，0：按 ID 倒序，1：按 updated_at 倒序，默认按 ID 倒序",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "项目状态（位掩码），用于复合查询，默认不筛选查询",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "用户 ID，默认不筛选用户",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "项目所属的作品集 ID，默认不筛选作品集",
                        "name": "workset_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dtos.ProjectBasic"
                            }
                        }
                    },
                    "400": {
                        "description": "无效的请求参数",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "创建一个新的项目",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project_proc"
                ],
                "summary": "创建项目",
                "parameters": [
                    {
                        "description": "创建项目的请求体",
                        "name": "body_params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateProjectRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.ProjectCreatedInfo"
                        }
                    },
                    "400": {
                        "description": "无效的请求参数",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/projects/{id}": {
            "get": {
                "description": "获取指定项目的详细信息，包括翻译、校对、嵌字、审核、发布等状态",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project"
                ],
                "summary": "获取项目详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "项目 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.ProjectDetail"
                        }
                    },
                    "400": {
                        "description": "无效的请求参数",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除指定的项目，需提供项目 ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project_proc"
                ],
                "summary": "删除项目",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "项目 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/handlers.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "无效的请求参数",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "更新指定的项目，需提供项目 ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project_proc"
                ],
                "summary": "更新项目",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "项目 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新的项目信息",
                        "name": "body_params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UpdateProjectRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/handlers.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "无效的请求参数",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/users/{id}/projects": {
            "get": {
                "description": "获取指定用户参与的所有项目列表，支持分页和复合查询",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "获取用户参与的项目列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码，默认值为 1",
                        "name": "page_serial",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量，默认值为 10",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "项目状态（位掩码），用于复合查询，默认不筛选查询",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "项目所属的作品集 ID，默认不筛选作品集",
                        "name": "workset_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "用户 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dtos.ProjectBasic"
                            }
                        }
                    },
                    "400": {
                        "description": "无效的请求参数",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dtos.CreateProjectRequest": {
            "type": "object",
            "required": [
                "title",
                "workset_id"
            ],
            "properties": {
                "allow_auto_join": {
                    "description": "是否允许自动加入，默认为 false",
                    "type": "boolean"
                },
                "description": {
                    "description": "项目描述",
                    "type": "string"
                },
                "is_hidden": {
                    "description": "是否隐藏项目，默认为 false",
                    "type": "boolean"
                },
                "title": {
                    "description": "项目标题，必填",
                    "type": "string"
                },
                "workset_id": {
                    "description": "作品集 ID，必填",
                    "type": "integer"
                }
            }
        },
...
        "dtos.MemberLabor": {
            "type": "object",
            "properties": {
                "joined_time": {
                    "description": "加入的",
                    "type": "string"
                },
                "labor_role": {
                    "description": "成员在项目中的角色，使用位掩码表示",
                    "type": "integer"
                },
                "nickname": {
                    "description": "昵称",
                    "type": "string"
                },
                "user_id": {
                    "description": "用户 ID",
                    "type": "integer"
                }
            }
        },
...
        "dtos.ProjectBasic": {
            "type": "object",
            "properties": {
                "allow_auto_join": {
                    "description": "是否允许自动加入",
                    "type": "boolean"
                },
                "id": {
                    "description": "项目 ID",
                    "type": "integer"
                },
                "is_hidden": {
                    "description": "是否是隐藏项目",
                    "type": "boolean"
                },
                "is_published": {
                    "description": "是否已发布",
                    "type": "boolean"
                },
                "labors": {
                    "description": "补充的成员分工信息",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.MemberLabor"
                    }
                },
                "legacy_id": {
                    "description": "历史遗留序号",
                    "type": "integer"
                },
                "moetran_id": {
                    "description": "龙译 ID",
                    "type": "string"
                },
                "status": {
                    "description": "项目状态，使用位掩码表示",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dtos.ProjectOverallStatus"
                        }
                    ]
                },
                "title": {
                    "description": "项目名称",
                    "type": "string"
                },
                "workset_id": {
                    "description": "所属作品集 ID",
                    "type": "integer"
                },
                "workset_index": {
                    "description": "作品集内的序号",
                    "type": "integer"
                }
            }
        },
        "dtos.ProjectCreatedInfo": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "响应消息",
                    "type": "string"
                },
                "moetran_id": {
                    "description": "龙译项目 ID",
                    "type": "string"
                },
                "project_id": {
                    "description": "创建的项目 ID",
                    "type": "integer"
                }
            }
        },
        "dtos.ProjectDetail": {
            "type": "object",
            "properties": {
                "allow_auto_join": {
                    "description": "是否允许自动加入",
                    "type": "boolean"
                },
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "description": {
                    "description": "项目描述",
                    "type": "string"
                },
                "id": {
                    "description": "项目 ID",
                    "type": "integer"
                },
                "is_hidden": {
                    "description": "是否是隐藏项目",
                    "type": "boolean"
                },
                "is_published": {
                    "description": "是否已发布",
                    "type": "boolean"
                },
                "labors": {
                    "description": "补充的成员分工信息",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.MemberLabor"
                    }
                },
                "legacy_id": {
                    "description": "历史遗留序号",
                    "type": "integer"
                },
                "moetran_id": {
                    "description": "龙译 ID",
                    "type": "string"
                },
                "status": {
                    "description": "项目状态，使用位掩码表示",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dtos.ProjectOverallStatus"
                        }
                    ]
                },
                "title": {
                    "description": "项目名称",
                    "type": "string"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                },
                "workset_id": {
                    "description": "所属作品集 ID",
                    "type": "integer"
                },
                "workset_index": {
                    "description": "作品集内的序号",
                    "type": "integer"
                }
            }
        },
        "dtos.ProjectOverallStatus": {
            "type": "integer",
            "enum": [
                3,
                12,
                48,
                192,
                768
            ],
            "x-enum-comments": {
                "PROJECT_STATUS_LETTER_MASK": "二进制 48",
                "PROJECT_STATUS_PROOF_MASK": "二进制 12",
                "PROJECT_STATUS_PUBLISH_MASK": "二进制 768",
                "PROJECT_STATUS_REVIEW_MASK": "二进制 192",
                "PROJECT_STATUS_TRANSLATE_MASK": "二进制 3"
            },
            "x-enum-varnames": [
                "PROJECT_STATUS_TRANSLATE_MASK",
                "PROJECT_STATUS_PROOF_MASK",
                "PROJECT_STATUS_LETTER_MASK",
                "PROJECT_STATUS_REVIEW_MASK",
                "PROJECT_STATUS_PUBLISH_MASK"
            ]
        },
...
        "dtos.UpdateProjectRequest": {
            "type": "object",
            "required": [
                "project_id",
                "title"
            ],
            "properties": {
                "description": {
                    "description": "项目描述",
                    "type": "string"
                },
                "project_id": {
                    "description": "项目 ID，必填",
                    "type": "integer"
                },
                "status": {
                    "description": "项目状态，使用位掩码表示",
                    "type": "integer"
                },
                "title": {
                    "description": "项目标题",
                    "type": "string"
                }
            }
        },
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "description": "可选，在 service 层发生错误时提供详细信息",
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "handlers.SuccessResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "description": "可选，提供额外的成功信息"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}